+++ src/app/(home)/@modal/(.)onboarding/page.tsx
@@ -1,4 +1,4 @@
-import OnboardingModal from "@/app/_components/OnboardingModal";
+import OnboardingModal from "@/app/_components/OnboardingModal/OnboardingModal";
 
 export default function Modal() {
   return <OnboardingModal />;
