+++ src/app/(home)/_components/ContentSlider/ContentSlider.module.css
