+++ src/app/(home)/_components/IndexIndicator/IndexIndicator.tsx
@@ -1,6 +1,6 @@
-import Number1 from "@/../public/assets/number_1.svg";
-import Number2 from "@/../public/assets/number_2.svg";
-import Number3 from "@/../public/assets/number_3.svg";
+import Number1 from "../../../../../public/assets/number_1.svg";
+import Number2 from "../../../../../public/assets/number_2.svg";
+import Number3 from "../../../../../public/assets/number_3.svg";
 import Image from "next/image";
 
 export default function IndexIndicator({ index }: { index: number }) {
