+++ src/app/(home)/loading.tsx
@@ -1,5 +1,5 @@
-import { categoryFallback } from "../_components/CategoryFallback";
-import ContentFallBack from "./_components/ContentSliderFallback";
+import { categoryFallback } from "../_components/CategoryFallBack/CategoryFallback";
+import ContentFallBack from "./_components/ContentSliderFallback/ContentSliderFallback";
 import styles from "./loading.module.css";
 export default function Loading() {
   const CategoryBar = () => categoryFallback;
