+++ src/app/_components/BottomTabBar/BottomTabBar.module.css
