+++ src/app/content/page.tsx
@@ -1,8 +1,8 @@
 import { Metadata } from "next";
-import ContentList from "./_components/ContentList";
+import ContentList from "./_components/ContentList/ContentList";
 import styles from "./page.module.css";
 import { Suspense } from "react";
-import { contentListFallBack } from "./_components/ContentListFallback";
+import { contentListFallBack } from "./_components/ContentListFallback/ContentListFallback";
 
 export const metadata: Metadata = {
   title: "Dev a Nice Day | 콘텐츠 목록",
