+++ src/app/onboarding/page.tsx
@@ -1,4 +1,4 @@
-import OnboardingModal from "../_components/OnboardingModal";
+import OnboardingModal from "../_components/OnboardingModal/OnboardingModal";
 import styles from "./page.module.css";
 
 export default function Onboarding() {
